{"schema":"decodex/commit/1","summary":"Modularize authority and state evidence surfaces","authority":"manual"}
